+++ VirtualConsole/Device.h
@@ -8,8 +8,8 @@ public:
 
     virtual ~Device() = default;
 
-    virtual uint8_t read(uint16_t address) = 0;
-    virtual void write(uint16_t address, uint8_t value) = 0;
+    virtual uint8_t read(uint32_t address) = 0;
+    virtual void write(uint32_t address, uint8_t value) = 0;
 
     virtual void tick() {}
